refactor(admin): move CRUD stores to runes
